Create debts-429c5441.py
https://www.hackerearth.com/practice/basic-programming/implementation/basics-of-implementation/practice-problems/algorithm/debts-429c5441/description/


There are  people living in a neighborhood. When in debt, neighbors borrow money from each other to clear their debts. A neighbor has already borrowed money from another neighbor for  times to clear a debt. 

All the neighbors want to clear what they owe each other. Their plan is to clear their debts in such a way that the total number of transactions is minimized because the fee per transaction is very high. For example, if the  person pays the  person  dollars, then the amount of this particular transaction is . 

You are required to minimize the sum of the transaction amount.

Input format

First line: Two integers  and 
Next  lines: Three integers , , and  which means that the  person has lent the  person  amount of dollars
Output format

Print only one integer that represents the minimum sum of the transaction amount.

Constraints




It is guaranteed that  and  are distinct.

Sample input

2 3
1 2 10
1 2 3
2 1 5

Sample output

8