Create power-of-four.py
https://www.hackerearth.com/practice/basic-programming/bit-manipulation/basics-of-bit-manipulation/practice-problems/algorithm/power-of-four/description/

Due to Corona many countries are facing shortage of medical facilities as a result of it China agreed to render its help to that country who is able to solve this problem.

Problem goes like : China gave a number N ,you have to find the smallest number X greater than or equal to N.

The number X is correct if it can be represented as sum of distinct powers of 4.

Input

The first line of the input contains one integer t (1≤t≤1000) — the number of testcases. 

The only line of the query contains one integer n (1≤n≤100000).

Output

For each query, print such smallest integer x (where n≤x) that is a solution.

Note: Copy and paste is disabled for all the questions. Try not to use other ides.

SAMPLE INPUT 
2
5
6
SAMPLE OUTPUT 
5
16
Explanation
For N =5 

We can represent it as 41+40 , which is distinct.

For N=6

We can represent it as 41+40+40, which is not distinct next number can b formed is 16 i.e 42